chore: move component props to /common
