+++ src/classic/components/Hero/Hero.tsx
@@ -1,12 +1,4 @@
-export interface HeroProps {
-  sectionIdx?: number
-  logoUrl?: string
-  heroTitle?: string
-  heroCaption?: string
-  buttonLabel?: string
-  buttonUrl?: string
-  bgUrl?: string
-}
+import { HeroProps } from "~/common"
 
 export default function Hero({
   heroTitle,
