+++ src/classic/components/Infobar/Infobar.tsx
@@ -1,14 +1,6 @@
 import { ArrowRightIcon } from "@heroicons/react/24/outline"
 import { HomepageSectionWrapper } from "../HomepageSectionWrapper"
-
-export interface InfobarProps {
-  sectionIdx?: number
-  title?: string
-  subtitle?: string
-  description?: string
-  buttonLabel?: string
-  buttonUrl?: string
-}
+import { InfobarProps } from "~/common"
 
 const Infobar = ({
   sectionIdx,
