+++ src/classic/components/index.ts
@@ -1,12 +1,14 @@
 export { default as Button } from "./Button"
-export { default as Footer } from "./Footer"
-export { default as Hero } from "./Hero"
 export { default as Cards } from "./Cards"
-export { default as Infobar } from "./Infobar"
-export { default as InfoPic } from "./Infopic"
 export { default as Content } from "./Content"
-export { default as Navbar } from "./Navbar"
+export { default as Footer } from "./Footer"
 export { default as Header } from "./Header"
+export { default as Hero } from "./Hero"
+export { default as Image } from "./Image"
+export { default as Infobar } from "./Infobar"
+export { default as InfoCards } from "./InfoCards"
+export { default as InfoPic } from "./Infopic"
+export { default as Masthead } from "./Masthead"
+export { default as Navbar } from "./Navbar"
 export { default as Search } from "./Search"
 export { default as SidePane } from "./SidePane"
-export { default as Image } from "./Image"
