+++ src/common/Button.tsx
@@ -1,0 +1,5 @@
+export interface ButtonProps {
+  label: string
+}
+
+export default ButtonProps
