+++ src/common/Cards.tsx
@@ -1,0 +1,28 @@
+export interface SingleCard {
+  id: number
+  title: string
+  href: string
+  description: string
+  imageUrl: string
+  date: string
+  datetime: string
+  category?: {
+    title: string
+    href: string
+  }
+  author: {
+    name: string
+    role: string
+    href: string
+    imageUrl: string
+  }
+}
+
+export interface CardsProps {
+  sectionIdx?: number
+  sectionTitle: string
+  sectionCaption: string
+  cards: SingleCard[]
+}
+
+export default CardsProps
