+++ src/common/Content.tsx
@@ -1,0 +1,5 @@
+export interface ContentProps {
+  markdown: string
+}
+
+export default ContentProps
