+++ src/common/Header.tsx
@@ -1,0 +1,8 @@
+import { Sitemap } from "~/engine/render"
+
+export interface HeaderProps {
+  permalink: string
+  sitemap: Sitemap
+}
+
+export default HeaderProps
