+++ src/common/Hero.tsx
@@ -1,0 +1,11 @@
+export interface HeroProps {
+  sectionIdx?: number
+  logoUrl?: string
+  heroTitle?: string
+  heroCaption?: string
+  buttonLabel?: string
+  buttonUrl?: string
+  bgUrl?: string
+}
+
+export default HeroProps
