+++ src/common/Image.tsx
@@ -1,0 +1,9 @@
+export interface ImageProps {
+  src: string
+  alt: string
+  width?: number
+  href?: string
+  openInNewTab?: boolean
+}
+
+export default ImageProps
