+++ src/common/Infobar.tsx
@@ -1,0 +1,10 @@
+export interface InfobarProps {
+  sectionIdx?: number
+  title?: string
+  subtitle?: string
+  description?: string
+  buttonLabel?: string
+  buttonUrl?: string
+}
+
+export default InfobarProps
