+++ src/common/Infopic.tsx
@@ -1,0 +1,13 @@
+export interface InfopicProps {
+  sectionIndex?: number
+  title: string
+  subtitle?: string
+  description?: string
+  imageAlt?: string
+  imageSrc: string
+  buttonLabel?: string
+  buttonUrl?: string
+  isLeftVariant?: boolean
+}
+
+export default InfopicProps
