+++ src/common/Masthead.tsx
@@ -1,0 +1,5 @@
+export interface MastheadProps {
+  isStaging?: boolean
+}
+
+export default MastheadProps
