+++ src/common/Navbar.tsx
@@ -1,0 +1,20 @@
+export interface NavbarLink {
+  type: "single" | "dropdown"
+  name: string
+  eventKey?: string
+  url?: string
+  links?: NavbarLink[]
+}
+
+export interface IsomerNavProps {
+  id?: string
+  logo: { url: string; alt: string }
+
+  links: NavbarLink[]
+  search?: {
+    isEnabled: boolean
+    searchUrl?: string
+  }
+}
+
+export default IsomerNavProps
