+++ src/common/Search.tsx
@@ -1,0 +1,12 @@
+export type SearchRecord = {
+  id: string
+  title: string
+  content: string
+  url: string
+}
+
+export interface SearchProps {
+  index: SearchRecord[]
+}
+
+export default SearchProps
