+++ src/common/SidePane.tsx
@@ -1,0 +1,8 @@
+import { Sitemap } from "~/engine/render"
+
+export interface SidePaneProps {
+  sitemap: Sitemap
+  currentPermalink: string
+}
+
+export default SidePaneProps
