+++ src/common/index.ts
@@ -1,0 +1,14 @@
+export { default as ButtonProps } from "./Button"
+export { default as CardsProps } from "./Cards"
+export { default as ContentProps } from "./Content"
+export { default as FooterProps } from "./Footer"
+export { default as HeaderProps } from "./Header"
+export { default as HeroProps } from "./Hero"
+export { default as ImageProps } from "./Image"
+export { default as InfobarProps } from "./Infobar"
+export { default as InfoCardsProps } from "./InfoCards"
+export { default as InfopicProps } from "./Infopic"
+export { default as MastheadProps } from "./Masthead"
+export { default as NavbarProps } from "./Navbar"
+export { default as SearchProps } from "./Search"
+export { default as SidePaneProps } from "./SidePane"
